+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- CommandDeps as the AbstractMultitronCommand's only dependency
+- MultitronFactory
 
 
 ## [1.0.0-beta5] - 2025-06-24